+++ collect_app/src/main/java/org/odk/collect/android/widgets/range/RangeIntegerWidget.kt
@@ -50,7 +50,7 @@ class RangeIntegerWidget(
                         requestDisallowInterceptTouchEvent(it)
                     },
                     onValueChange = {
-                        rangeSliderState = rangeSliderState.copy(sliderValue = it)
+                        rangeSliderState = rangeSliderState.copy(sliderValue = it.toBigDecimal())
                     },
                     onValueChangeFinished = {
                         widgetValueChanged()
