+++ collect_app/src/test/java/org/odk/collect/android/widgets/range/RangeSliderStateTest.kt
@@ -47,8 +47,8 @@ class RangeSliderStateTest {
 
         val state = RangeSliderState.fromPrompt(prompt)
 
-        assertThat(state.sliderValue, equalTo(0.7f))
-        assertThat(state.realValue, equalTo(7.0))
+        assertThat(state.sliderValue, equalTo(0.7.toBigDecimal()))
+        assertThat(state.realValue, equalTo(7.0.toBigDecimal()))
     }
 
     @Test
@@ -62,8 +62,8 @@ class RangeSliderStateTest {
 
         val state = RangeSliderState.fromPrompt(prompt)
 
-        assertThat(state.sliderValue, equalTo(0.3f))
-        assertThat(state.realValue, equalTo(7.0))
+        assertThat(state.sliderValue, equalTo(0.3.toBigDecimal()))
+        assertThat(state.realValue, equalTo(7.0.toBigDecimal()))
     }
 
     @Test
@@ -90,10 +90,10 @@ class RangeSliderStateTest {
 
         val state = RangeSliderState.fromPrompt(prompt)
 
-        assertThat(state.rangeStart, equalTo(0f))
+        assertThat(state.rangeStart, equalTo(0.0.toBigDecimal()))
         assertThat(state.startLabel, equalTo("0"))
 
-        assertThat(state.rangeEnd, equalTo(10f))
+        assertThat(state.rangeEnd, equalTo(10.0.toBigDecimal()))
         assertThat(state.endLabel, equalTo("10"))
     }
 
@@ -107,10 +107,10 @@ class RangeSliderStateTest {
 
         val state = RangeSliderState.fromPrompt(prompt)
 
-        assertThat(state.rangeStart, equalTo(10f))
+        assertThat(state.rangeStart, equalTo(10.0.toBigDecimal()))
         assertThat(state.startLabel, equalTo("10"))
 
-        assertThat(state.rangeEnd, equalTo(0f))
+        assertThat(state.rangeEnd, equalTo(0.0.toBigDecimal()))
         assertThat(state.endLabel, equalTo("0"))
     }
 
@@ -409,10 +409,10 @@ class RangeSliderStateTest {
     @Test
     fun `realValue rounds correctly to nearest step for ascending range with integer step`() {
         val state = RangeSliderState(
-            sliderValue = 0.34f,
-            rangeStart = 0f,
-            rangeEnd = 10f,
-            step = 1f,
+            sliderValue = 0.34.toBigDecimal(),
+            rangeStart = 0.toBigDecimal(),
+            rangeEnd = 10.toBigDecimal(),
+            step = 1.toBigDecimal(),
             numOfSteps = 9,
             isDiscrete = true,
             isHorizontal = true,
@@ -421,16 +421,16 @@ class RangeSliderStateTest {
             numOfTicks = 11
         )
 
-        assertThat(state.realValue, equalTo(3.0))
+        assertThat(state.realValue, equalTo(3.toBigDecimal()))
     }
 
     @Test
     fun `realValue rounds correctly to nearest step for descending range with integer step`() {
         val state = RangeSliderState(
-            sliderValue = 0.34f,
-            rangeStart = 10f,
-            rangeEnd = 0f,
-            step = 1f,
+            sliderValue = 0.34.toBigDecimal(),
+            rangeStart = 10.toBigDecimal(),
+            rangeEnd = 0.toBigDecimal(),
+            step = 1.toBigDecimal(),
             numOfSteps = 9,
             isDiscrete = true,
             isHorizontal = true,
@@ -439,16 +439,16 @@ class RangeSliderStateTest {
             numOfTicks = 11
         )
 
-        assertThat(state.realValue, equalTo(7.0))
+        assertThat(state.realValue, equalTo(7.toBigDecimal()))
     }
 
     @Test
     fun `realValue rounds correctly to nearest step for ascending range with decimal step`() {
         val state = RangeSliderState(
-            sliderValue = 0.14444445f,
-            rangeStart = 1f,
-            rangeEnd = 10f,
-            step = 0.1f,
+            sliderValue = 0.14444445.toBigDecimal(),
+            rangeStart = 1.toBigDecimal(),
+            rangeEnd = 10.toBigDecimal(),
+            step = 0.1.toBigDecimal(),
             numOfSteps = 89,
             isDiscrete = true,
             isHorizontal = true,
@@ -457,16 +457,16 @@ class RangeSliderStateTest {
             numOfTicks = 11
         )
 
-        assertThat(state.realValue, equalTo(2.3))
+        assertThat(state.realValue, equalTo(2.3.toBigDecimal()))
     }
 
     @Test
     fun `realValue rounds correctly to nearest step for descending range with decimal step`() {
         val state = RangeSliderState(
-            sliderValue = 0.14444445f,
-            rangeStart = 10f,
-            rangeEnd = 1f,
-            step = 0.1f,
+            sliderValue = 0.14444445.toBigDecimal(),
+            rangeStart = 10.toBigDecimal(),
+            rangeEnd = 1.toBigDecimal(),
+            step = 0.1.toBigDecimal(),
             numOfSteps = 89,
             isDiscrete = true,
             isHorizontal = true,
@@ -475,7 +475,7 @@ class RangeSliderStateTest {
             numOfTicks = 11
         )
 
-        assertThat(state.realValue, equalTo(8.7))
+        assertThat(state.realValue, equalTo(8.7.toBigDecimal()))
     }
 
     private fun mockQuestion(rangeStart: Float, rangeEnd: Float, rangeStep: Float) {
